backend/requirements.txt
sqlalchemy
psycopg2-binary
numpy
lapjv
sklearn
scipy
tensorflow>=2
Pillow
boto3